Extracted header file from SteeredWheelBaseController.
